Shortcut through simplification step if distanceTolerance is negative
If the simplification tolerance has been set to less than 0.0 then there's
no need to run the DouglasPeuckerSimplifier algorithm. This change
checks the tolerance and simply returns a copy of the Geometry if no
simplification is requested.
This change is required for the case where the user wants no
simplification to take place e.g. to preserve all intersection nodes.
Without this change any negative value for distanceTolerance is squared
inside DPTransformer and immediately loses its meaning.